Renames
Rename all things to lowercase.
Rename classes back to original.
