+++ main.cpp
@@ -1,10 +1,10 @@
-#include "gameManager.h"
+#include "gamemanager.h"
 
 // All we need to do here is call the GameManager constructor, 
 //  since this will execute the game loop 
 
 int main()
 {
-    gameManager *game = new gameManager(); 
+    GameManager *game = new GameManager(); 
     return 0;
 }
