+++ objects/character.h
