+++ objects/mainchar.h
